Проверка на валидность видео

+++ lib/tasks/slides.rake
@@ -13,6 +13,9 @@ namespace :slides do
       .each do |slide|
         mp4_video_path = Rails.root.join("tmp/video/#{slide.video.key}.mp4")
         video = FFMPEG::Movie.new(slide.video.service_url)
+
+        next unless video.valid?
+
         File.delete(mp4_video_path) if File.exist?(mp4_video_path)
         video.transcode(
           mp4_video_path.to_s,
